requirements/requirements-base.txt
anyio==3.6.2
dependency-injector==4.39.1

fastapi==0.78.0

numpy==1.23.0

omegaconf==2.2.2
onnxruntime==1.19.0
opencv-python==4.6.0.66

opentelemetry-api==1.24.0
opentelemetry-exporter-otlp==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-sdk==1.24.0

prometheus-client==0.22.1
python-multipart==0.0.5
requests==2.32.3
uvicorn==0.18.2
